% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{applyQC}
\alias{applyQC}
\title{Use quality-control flags to alter argoFloats data}
\usage{
applyQC(x, flags = NULL, actions = NULL, debug = 0)
}
\arguments{
\item{x}{An \linkS4class{argoFloats} object of type \code{"argos"}, as created by
\code{\link[=readProfiles]{readProfiles()}}.}

\item{flags}{A vector of integers that are the quality-control flag
values for data that are to be set to \code{NA} in the returned result.
The default, \code{NULL}, means to use the flags set up by \code{\link[=readProfiles]{readProfiles()}}.
That means to regard as suspicious any data entries that are flagged
with QC codes of 0 (for data that have not yet been assessed),
3 (for "probably bad" data),
4 (for "bad" data),
6 (an unused flag),
7 (an unused flag), or
9 (for "missing" data).}

\item{actions}{The actions to perform. The default, \code{NULL}, means to
use the actions set up by \code{\link[=readProfiles]{readProfiles()}}, which, by default.
causes any data flagged as suspicious to be set to \code{NA}.}

\item{debug}{Passed to \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}.}
}
\value{
A copy of \code{x} but with each of the objects within its
\code{data} slot having been passed through \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}.
}
\description{
This function examines the quality-control (QC) flags within an \linkS4class{argoFloats}
object that was created by \code{\link[=readProfiles]{readProfiles()}}.  By default, it replaces all suspicious
data with \code{NA} values, so they will not appear in plots or be considered in calculations.
This is an important early step in processing, because suspicious argo floats commonly
report data that are suspicious based on statistical and physical measures, as
is illustrated in the \dQuote{Examples} section.
}
\details{
The work is done by using \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}
on each of the profiles stored within the object. In most cases, only
the object needs to be specified, for the default actions coincide with
common conventions for flags in argo data.
}
\examples{
library(argoFloats)
# Contrast TS diagrams for raw and flag-handled data
data(index)
i <- subset(index, 1:5) # first 5 profiles
raw <- readProfiles(getProfiles(i))
clean <- applyQC(raw)
par(mfrow=c(1, 2))
plot(raw, which="TS")
plot(clean, which="TS")

}
\references{
Carval, Thierry, Bob Keeley, Yasushi Takatsuki, Takashi Yoshida, Stephen Loch Loch,
Claudia Schmid, and Roger Goldsmith. Argo User’s Manual V3.3. Ifremer, 2019.
\url{https://doi.org/10.13155/29825}.
}
\author{
Dan Kelley
}
