% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{argoDefaultDestdir}
\alias{argoDefaultDestdir}
\alias{argoDefaultIndexAge}
\alias{argoDefaultProfileAge}
\title{Get the default Argo cache directory}
\usage{
argoDefaultDestdir()

argoDefaultIndexAge()

argoDefaultProfileAge()
}
\value{
A character value containing a path to the default Argo data directory.
}
\description{
By default, files downloaded from the Argo mirror are placed in
\verb{~/data/argo}. You can configure where these files are stored by setting
\code{options(argoFloat.destdir="path/to/argo")}. To persist this value
between sessions you can place this value in your \code{.Rprofile}.
}
\examples{
argoDefaultDestdir()

}
