% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{[[,argoFloats-method}
\alias{[[,argoFloats-method}
\title{Look up a Value Within an argoFloats Object}
\usage{
\S4method{[[}{argoFloats}(x, i, j, ...)
}
\arguments{
\item{x}{an \linkS4class{argoFloats} object.}

\item{i}{a character value that specifies the item to be looked up;
see \dQuote{Details}.}

\item{j}{ignored.}

\item{...}{ignored.}
}
\description{
This function provides an easy way to look up values within an \linkS4class{argoFloats}
object, without the need to know the exact structure of the data. There are three
modes of operation, as determined by the \code{i} argument.
}
\details{
The possibilities for the value of \code{i} are as follows.
\enumerate{
\item If \code{i=="metadata"} then the \code{metadata} slot is returned.
\item If \code{i=="data"} then the \code{data} slot is returned.
\item If \code{i=="ftpRoot"} then the value of \code{ftpRoot} in the \code{metadata}
slot is returned.  This is used by \code{\link[=getProfiles]{getProfiles()}} to construct
a vector of URLs to download.
\item If \code{i=="ftpRoot"} then the value of \code{ftpRoot} in the \code{metadata}
\item If \code{i=="destdir"} then the \code{destdir} item in the \code{metadata} slot is
returned. This is used by \code{\link[=getProfiles]{getProfiles()}} to decide where to
save downloaded files, and later by \code{\link[=readProfiles]{readProfiles()}}, which
works with the output from \code{\link[=getProfiles]{getProfiles()}}.
\item If \code{i=="type"} then the \code{type} item in the \code{metadata} slot is
returned. This is \code{"index"} if the object was created with
\code{\link[=getIndex]{getIndex()}} and \code{"profiles"} if it was created with \code{\link[=getProfiles]{getProfiles()}}.
\item If \code{i=="profile"} and the object \code{type} is \code{argos}, as will
be the case if \code{x} is a return value from \code{\link[=readProfiles]{readProfiles()}}, return
either a list containing all profiles, or a requested profile;
see \code{\link[=readProfiles]{readProfiles()}} for examples.
\item Otherwise, \code{i} must be the name of an item in the \code{index} item
in the \code{data} slot, or a string that is made up of enough characters
to uniquely identify such an item, e.g. \code{"lon"} may be used as a
shortcut for \code{"longitude"}.
}
}
\author{
Dan Kelley
}
