% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{package}
\name{argoFloats-package}
\alias{argoFloats-package}
\title{A Package for Collections of Argo Float Profiles}
\description{
This package \strong{FIXME(jlh) Please write a few paragraphs here, after
discussions with dek and cgr.  The idea will be to get a clean
introductory paragraph that will organize our intentions for
the object structure.  Details ought to be added later, as
new things get written.  The oce package may provide a useful
guide.}
}
