% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsClearCache}
\alias{argoFloatsClearCache}
\title{Clear the Cache}
\usage{
argoFloatsClearCache(debug = 0L)
}
\arguments{
\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\value{
integer, returned invisibly, indicating the number of items removed.
}
\description{
Clear the cache.  This is meant mainly for developers working in interactive
sessions to test coding changes.
}
\seealso{
Other functions relating to cached values: 
\code{\link{argoFloatsGetFromCache}()},
\code{\link{argoFloatsIsCached}()},
\code{\link{argoFloatsListCache}()},
\code{\link{argoFloatsStoreInCache}()},
\code{\link{argoFloatsWhenCached}()}
}
\author{
Dan Kelley, making a thin copy of code written by Dewey Dunnington
}
\concept{functions relating to cached values}
