% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{argoFloatsDebug}
\alias{argoFloatsDebug}
\title{Possibly print debugging information.}
\usage{
argoFloatsDebug(debug = 0, ..., style = "plain", unindent = 0)
}
\arguments{
\item{debug}{an integer specifying the level of debugging. Values greater
than zero indicate that some printing should be done. Many functions}

\item{...}{content to be printed, analogous to the the ... argument
list of \code{\link[=cat]{cat()}}.}

\item{style}{character value indicating special formatting, with \code{"plain"}
for normal text, \code{"bold"} for bold-faced text, \code{"italic"} for italicized
text, \code{"red"} for red text, \code{"green"} for green text, or \code{"blue"} for blue
text. These codes may not be combined.}

\item{unindent}{integer specifying the degree of reverse indentation
to be done, as explained in the \dQuote{Details} sction.}
}
\description{
This function is intended mainly for use within the package, but users may
also call it directly in their own code.  Within the package, the value
of \code{debug} is generally reduced by 1 on each nested function call, leading
to indented messages. Most functions start and end with a call to
\code{\link[=argoFloatsDebug]{argoFloatsDebug()}} that has \code{style="bold"} and \code{unindent=1}.
}
\author{
Dan Kelley
}
