% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsIsCached}
\alias{argoFloatsIsCached}
\title{Check Whether an Item is Cached}
\usage{
argoFloatsIsCached(name, debug = 0)
}
\arguments{
\item{name}{character value, naming the item.}

\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\description{
Check Whether an Item is Cached
}
\author{
Dan Kelley, making a thin copy of code written by Dewey Dunnington
}
