% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsListCache}
\alias{argoFloatsListCache}
\title{List Items in the Cache}
\usage{
argoFloatsListCache(debug = 0L)
}
\arguments{
\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\value{
character vector of names of items.
}
\description{
List items in the cache.
}
\seealso{
Other functions relating to cached values: 
\code{\link{argoFloatsClearCache}()},
\code{\link{argoFloatsGetFromCache}()},
\code{\link{argoFloatsIsCached}()},
\code{\link{argoFloatsStoreInCache}()},
\code{\link{argoFloatsWhenCached}()}
}
\author{
Dan Kelley, making a thin copy of code written by Dewey Dunnington
}
\concept{functions relating to cached values}
