% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsStoreInCache}
\alias{argoFloatsStoreInCache}
\title{Store an item in the cache}
\usage{
argoFloatsStoreInCache(name, value, debug = 0)
}
\arguments{
\item{name}{character value, naming the item.}

\item{value}{the new contents of the item.}

\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\description{
Store an item in the cache
}
