% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsWhenCached}
\alias{argoFloatsWhenCached}
\title{Check When an Item was Cached}
\usage{
argoFloatsWhenCached(name, debug = 0L)
}
\arguments{
\item{name}{character value, naming the item.}

\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\value{
A logical value indicating whether a cached value is available.
}
\description{
Check When an Item was Cached
}
\seealso{
Other functions relating to cached values: 
\code{\link{argoFloatsClearCache}()},
\code{\link{argoFloatsGetFromCache}()},
\code{\link{argoFloatsIsCached}()},
\code{\link{argoFloatsListCache}()},
\code{\link{argoFloatsStoreInCache}()}
}
\author{
Dan Kelley, making a thin copy of code written by Dewey Dunnington
}
\concept{functions relating to cached values}
