% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadWithRetries}
\alias{downloadWithRetries}
\title{Download and Cache a Dataset}
\usage{
downloadWithRetries(
  url,
  destdir = ".",
  destfile = NULL,
  mode = "wb",
  quiet = FALSE,
  force = FALSE,
  retries = 3,
  debug = 0
)
}
\arguments{
\item{url}{character value giving the full URL of a file to be downloaded.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{destfile}{character value indicating the name of the file.
If not supplied, then the file name is constructed from the other
parameters of the function call, so that subsequent calls with
the same parameters will yield the same result; this is useful
for caching.}

\item{mode}{character value passed to \code{\link[curl:curl_download]{curl::curl_download()}},
controlling the mode used in writing the local file. The default
value, \code{mode="wb"}, means to set the write format to binary. This
is necessary on Windows machines, which otherwise will alter newline
characters, ruining file integrity for binary files (such as
netcdf files). The setting is not required on other machines, but
it makes sense to use this default in all instances.  See the
documentation for \code{\link[curl:curl_download]{curl::curl_download()}}.}

\item{quiet}{logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}},
that indicates whether to download silently.  Since some downloads
can be slow, the default value is \code{quiet=FALSE}.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{retries}{integer telling how many times to retry a download,
if the first attempt fails.}

\item{debug}{integer value indicating level of debuggin. If this
is less than 1, no debugging is done. Otherwise, some functions
will print debugging information.  If a function call fails, the
first step should be to rerun the function with \code{debug=1},
to see if the output suggests a problem in the call.}
}
\value{
String indicating the full pathname to the downloaded file.
}
\description{
General function for downloading and caching a dataset.
}
\author{
Dan Kelley
}
