% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getProfiles}
\alias{getProfiles}
\title{Get all Profiles Named in an argoFloats Index}
\usage{
getProfiles(
  index,
  destdir = ".",
  force = FALSE,
  retries = 3,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{index}{an \linkS4class{argoFloats} object of type \code{"index"}, as created
by \code{\link[=getIndex]{getIndex()}}.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{retries}{integer telling how many times to retry a download,
if the first attempt fails.}

\item{quiet}{logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}},
that indicates whether to download silently.  Since some downloads
can be slow, the default value is \code{quiet=FALSE}.}

\item{debug}{integer value indicating level of debugging. If this
is less than 1, no debugging is done. Otherwise, some functions
will print debugging information.  If a function call fails, the
first step should be to rerun the function with \code{debug=1},
to see if the output suggests a problem in the call.}
}
\value{
A character vector of filenames that have been downloaded, or
that were already present in th \code{destdir} directory.
}
\description{
Get all Profiles Named in an argoFloats Index
}
