% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{handleFlags,argoFloats-method}
\alias{handleFlags,argoFloats-method}
\title{Handle flags in an argoFloats object of type "argos"}
\usage{
\S4method{handleFlags}{argoFloats}(object, flags = NULL, actions = NULL, where = NULL, debug = 0)
}
\arguments{
\item{object}{A \linkS4class{argoFloats} object of type \code{"argos"}, as created by
\code{\link[=readProfiles]{readProfiles()}}.}

\item{flags}{A vector of integers that are flag values for data that are
to be set to \code{NA} in the returned result.  The default, \code{NULL}, means to
use the flags set up by \code{\link[=readProfiles]{readProfiles()}}.}

\item{actions}{The actions to perform. The default, \code{NULL}, means to
use the actions set up by \code{\link[=readProfiles]{readProfiles()}}.}

\item{where}{Passed to \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}.}

\item{debug}{Passed to \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}.}
}
\value{
A copy of \code{object} but with each argo object contained therein having
been passed through \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}.
}
\description{
The work is done by using \code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}} on each of the
argo profiles that are stored in the \code{argos} item within the \code{data} slot
of \code{object}.  Apart from \code{object}, all arguments are passed to
\code{\link[oce:handleFlags,argo-method]{oce::handleFlags,argo-method()}}.
}
\examples{
library(argoFloats)
# Contrast TS diagrams for raw and flag-handled data
data(index)
i <- subset(index, 1:5) # first 5 profiles
a <- readProfiles(getProfiles(i))
A <- handleFlags(a)
par(mfrow=c(1, 2))
plot(a, which="TS")
plot(A, which="TS")

}
\author{
Dan Kelley
}
