% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{data}
\name{index}
\alias{index}
\title{A sample index of profiles}
\description{
This is created by subsetting a global index to the 1783 Argo profiles
that were within a 300km radius of Sable Island as of 2020 March 14, using
the following code. Note the \code{destdir} value, which controls where a
subsequent \code{\link[=getProfiles]{getProfiles()}} call will save data files.
\preformatted{
library(oce)
indexAll <- getIndex(destdir="~/data/argo")
index <- subset(indexAll, circle=list(longitude=-59.91, latitude=43.93, radius=300))
}
}
\examples{
library(oce)
library(ocedata)
library(argoFloats)
data(coastlineWorldFine, package="ocedata")
data(topoWorld, package="oce")
data(index, package="argoFloats")
plot(coastlineWorldFine,
     clongitude=-59.91, clatitude=43.93, span=1000)
points(index[["longitude"]], index[["latitude"]], pch=20, col=rgb(1,0,0,0.1))
contour(topoWorld[["longitude"]], topoWorld[["latitude"]], -topoWorld[["z"]],
           add=TRUE, levels=c(1000, 2000), col="blue", lty=c(2,1))
legend("topleft", col="blue", lty=c(2,1), title="Water Depth [m]", legend=c("1000", "2000"),
       bg="white")

}
\concept{datasets provided with argoFloats}
