% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{data}
\name{indexMerged}
\alias{indexMerged}
\title{A sample index of merged argo and biogeochemical-argo profiles}
\description{
This is created by subsetting a global index to the 39 BGC Argo profiles
that were within a 300km radius of Marsh Harbour, Abaco Island,
Bahamas, as of 2020 March 23, using
the following code.
\preformatted{
library(argoFloats)
indexAll <- getIndex(filename="merged")
indexMerged <- subset(indexAll, circle=list(longitude=-77.06, latitude=26.54, radius=300))
save(indexMerged, file="indexMerged.rda")
tools::resaveRdaFiles('indexMerged.rda')
}
}
\examples{
library(argoFloats)
data(indexMerged)
plot(indexMerged)
summary(indexMerged)
unique(indexMerged[["parameters"]])

}
\seealso{
Other datasets provided with argoFloats: 
\code{\link{indexBgc}},
\code{\link{index}}
}
\concept{datasets provided with argoFloats}
