% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{data}
\name{indexSynthetic}
\alias{indexSynthetic}
\title{A sample index of synthetic (i.e. combined) argo and biogeochemical-argo profiles}
\description{
This was created by subsetting a global index to the 39 BGC Argo profiles
that were within a 300km radius of Marsh Harbour, Abaco Island,
Bahamas, as of 2020 May 15, using the following code.\preformatted{library(argoFloats)
indexAll <- getIndex("synthetic")
indexSynthetic <- subset(indexAll,
    circle=list(longitude=-77.06, latitude=26.54, radius=300))
}
}
\section{Historical note}{

This "synthetic" type of index is a replacement for the older "merged" index.  See
\url{http://www.argodatamgt.org/Data-Mgt-Team/News/BGC-Argo-synthetic-profiles-distributed-on-GDAC}
for more on the data file format, reasons for the change, and timetable for transition
from "merged".
}

\examples{
library(argoFloats)
data(indexSynthetic)
plot(indexSynthetic, bathymetry=FALSE)
summary(indexSynthetic)
unique(indexSynthetic[["parameters"]])
}
\seealso{
Other datasets provided with argoFloats: 
\code{\link{indexBgc}},
\code{\link{indexMerged}},
\code{\link{index}}
}
\concept{datasets provided with argoFloats}
