% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,argoFloats-method}
\alias{plot,argoFloats-method}
\title{Plot an argoFloats object}
\usage{
\S4method{plot}{argoFloats}(
  x,
  which = "map",
  xlab = NULL,
  ylab = NULL,
  cex = NULL,
  col = NULL,
  pch = NULL,
  mar = NULL,
  mgp = NULL,
  eos = "gsw",
  debug = 0,
  ...
)
}
\arguments{
\item{x}{an \code{\linkS4class{argoFloats}} object.}

\item{which}{a string that indicates the type of plot; see \dQuote{Details}.}

\item{xlab}{character value indicating the name for the horizontal axis, or
\code{NULL}, which indicates that this function should choose an appropriate name
depending on the value of \code{which}. Note that \code{xlab} is not obeyed if
\code{which="TS"}, because altering that label can be confusing to the user.}

\item{ylab}{as \code{xlab}, but for th vertical axis.}

\item{cex}{character expansion factor for plot symbols, or \code{NULL}, to get an
value that depends on the value of \code{which}.}

\item{col}{colour to be used for plot symbols, or \code{NULL}, to get an
value that depends on the value of \code{which}.}

\item{pch}{number indicating the type of plot symbol, or \code{NULL}, to get an
value that depends on the value of \code{which}.}

\item{mar}{either a four-element vector giving the margins to be used for
the plot (see \code{\link[=par]{par()}} for details), or \code{NULL}, which means to use
\link{par}\code{("mar")}.}

\item{mgp}{either a three-element vector giving the geometry for
axis labels (see \code{\link[=par]{par()}} for details), or \code{NULL}, which means to use
\link{par}\code{("mgp")}.}

\item{eos}{character value indicating the equation of state to use
if \code{which="TS"}.  This must be \code{"gsw"} (the default) or \code{"unesco"};
see \code{\link[oce:plotTS]{oce::plotTS()}}.}

\item{debug}{an integer specifying the level of debugging.}

\item{\dots}{extra arguments passed to the plot calls that are made
to within this function. A common use for \code{which="map"} cases is
to set \code{xlim} and \code{ylim} to get enlarge the plot region, so a coastline
becomes visible.}
}
\description{
The action depends on the \code{type} of the object, and
this is set up by the function that created the object;
see \dQuote{Details}.
}
\details{
The various plot types are as follows.
\itemize{
\item For \code{which="map"}, a map of profile locations is created. This
only works if the \code{type} is \code{"index"} (meaning that \code{x} was created
by \code{\link[=getIndex]{getIndex()}} or a subset of such an object, created with
\code{\link[=subset,argoFloats-method]{subset,argoFloats-method()}}), or \code{argos} (meaning that
\code{x} was created with \code{\link[=readProfiles]{readProfiles()}}.  The plot range is
auto-selected.  If the \code{ocedata} package
is available, then its \code{coastlineWorldFine} dataset is used to draw
a coastline (which will be visible only if the plot region
is large enough); otherwise, if the \code{oce} package is available, then its
\code{coastlineWorld} dataset is used.
\item For \code{which="TS"},  an overall TS plot is created.  This
only works if \code{x} is an object that was created by  \code{\link[=getProfiles]{getProfiles()}}.
The scales for the plot can be altered by putting \code{Slim} and \code{Tlim}
arguments in the \code{...} list; see the documentation for \code{\link[oce:plotTS]{oce::plotTS()}}
for other arguments that can be provided.
}
}
\examples{
# Example 1: map profiles in index, highlighting a neighborhood of 10
library(argoFloats)
library(oce)
data(index)
plot(index, which="map")
lon <- index[["longitude"]]
lat <- index[["latitude"]]
dist <- geodDist(lon, lat, -77.06, 26.54)
o <- order(dist)
index10 <- subset(index, o[1:10])
points(index10[["longitude"]], index10[["latitude"]], col=2)

# Example 2: TS of first 10 profiles in the index
\dontrun{
profiles10 <- getProfiles(index10, destdir="~/data/argo")
argos10 <- readProfiles(profiles10)
plot(argos10, which="TS")}

}
\author{
Dan Kelley
}
