% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcApp.R
\name{qcApp}
\alias{qcApp}
\title{Interactive app for analying Argo QC}
\usage{
qcApp(debug = 0)
}
\arguments{
\item{debug}{integer value that controls how much information \code{mapApp()} prints
to the console as it works.  The default value of 0 leads to a fairly limited
amount of printing, while higher values lead to more information. This information
can be helpful in diagnosing problems or bottlenecks.}
}
\description{
The GUI permits user to analyze QC of Argo data. As of 2020-02-05, this was still in early stages of development.
}
\examples{
if (interactive()) {
    library(argoFloats)
    qcApp()
}

}
\author{
Jaimie Harbin
}
