% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{R3901602_163.nc}
\alias{R3901602_163.nc}
\title{Sample Argo File (Real-Time Core Data)}
\description{
This is NetCDF file for real-time data for cycle 163 of Argo float 3901602, downloaded from
\code{https://data-argo.ifremer.fr/dac/coriolis/3901602/profiles/R3901602_163.nc}
on 2021 March 7.
}
\examples{
library(argoFloats)
a <- readProfiles(system.file("extdata", "R3901602_163.nc", package = "argoFloats"))
summary(a)
summary(a[[1]])

}
\seealso{
Other raw datasets: 
\code{\link{D4900785_048.nc}},
\code{\link{SD5903586_001.nc}},
\code{\link{SR2902204_131.nc}}
}
\concept{raw datasets}
