% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{SD5903586_001.nc}
\alias{SD5903586_001.nc}
\title{Sample argo file (delayed synthetic bgc data)}
\description{
This is profile 001 for float SD5903586, downloaded on June 20,
2020.  As its filename indicates, it holds "synthetic" data
in "delayed" mode. The oxygen values are adjusted by 16\%, as
is shown here, and in the documentation for \code{\link[=useAdjusted]{useAdjusted()}}.
}
\examples{
library(argoFloats)
a <- readProfiles(system.file("extdata", "SD5903586_001.nc", package="argoFloats"))
a1 <- a[[1]]
# Illustrate oxygen adjustment
p <- a1[["pressure"]]
O <- a1[["oxygen"]]
Oa  <- a1[["oxygenAdjusted"]]
Percent <- 100 * (Oa - O) / O
hist(Percent, main="Oxygen adjustment")

}
\seealso{
Other raw datasets: 
\code{\link{D4900785_048.nc}},
\code{\link{SR2902204_131.nc}}
}
\concept{raw datasets}
