% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,argoFloats-method}
\alias{show,argoFloats-method}
\title{Show information about argoFloats object}
\usage{
\S4method{show}{argoFloats}(object)
}
\arguments{
\item{object}{an \linkS4class{argoFloats} object.}
}
\description{
Show information about argoFloats object
}
\author{
Jaimie Harbin`
}
