% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showqc.R
\name{showQCTests}
\alias{showQCTests}
\title{Show which QC tests were performed and failed on an argos class}
\usage{
showQCTests(a)
}
\arguments{
\item{x}{a single float of argos class}
}
\value{
The associated QC test numbers performed and failed.
}
\description{
This function uses integer value from \code{\link[=hexToNibble]{hexToNibble()}} to indicate which
QC tests were performed and/or failed on an argos class object created by
\code{\link[=readProfiles]{readProfiles()}}
}
\examples{
\dontrun{
library(argoFloats)
data('index')
subset <- subset(index, 1)
profiles <- getProfiles(subset)
argos <- readProfiles(profiles)
argos1 <- argos[[1]]
showQCTests(argos[[1]])
  } 
}
