% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{SR2902204_131.nc}
\alias{SR2902204_131.nc}
\title{Sample argo file (synthetic realtime data)}
\description{
This is profile 131 for float 2902204, downloaded on June 8,
2020.  As its filename indicates, this is in the "synthetic"
data stream of "realtime" data.
}
\examples{
library(argoFloats)
a <- readProfiles(system.file("extdata", "SR2902204_131.nc", package="argoFloats"))
summary(a)
summary(a[[1]])

}
\seealso{
Other raw datasets: 
\code{\link{D4900785_048.nc}},
\code{\link{SD5903586_001.nc}}
}
\concept{raw datasets}
