% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{subset,argoFloats-method}
\alias{subset,argoFloats-method}
\alias{subset.argoFloats}
\title{Subset an argoFloats Object}
\usage{
\S4method{subset}{argoFloats}(x, subset, ...)
}
\arguments{
\item{x}{an \linkS4class{argoFloats} object.}

\item{subset}{optional numerical or logical vector that indicates which
indices of \code{x@data$index} to keep.  See example 1.}

\item{...}{a single named list. At the moment, the only possibility
is a list named \code{circle}, which holds elements named \code{longitude},
\code{latitude} and \code{radius}.  The first two specify a location, and}
}
\value{
An \linkS4class{argoFloats} object.
}
\description{
Return a subset of an \linkS4class{argoFloats} object, either by specifying
indices to keep (using the \code{subset} argument) or by specifying
a way to determine those indices (using the \code{...} argument).
}
\details{
At the moment, this method only works for \linkS4class{argoFloats} objects
of type \code{"index"}, as read by \code{\link[=getIndex]{getIndex()}}.
A future version will also handle \code{...} arguments named
\code{polygon}, \code{rectangle} and \code{time}.
}
\examples{
library(argoFloats)
\dontrun{
ai <- getIndex(file="argo_bio-profile_index.txt.gz", destdir="~/data/argo")
# Example 1: First three profiles in dataset.
aiFirstThree <- subset(ai, 1:3)
cat("First three longitudes:", paste(aiFirstThree[["longitude"]]), "\n")

# Example 2: Profiles within 50km of Bermuda
aiBermuda <- subset(ai, circle=list(longitude=-65.75, latitude=32.31, radius=50))
cat("Found", length(aiBermuda[["longitude"]]), "profiles near Bermuda\n")
}

}
\author{
Dan Kelley
}
