% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,argoFloats-method}
\alias{summary,argoFloats-method}
\title{Summarize an argoFloats Object}
\usage{
\S4method{summary}{argoFloats}(object, ...)
}
\arguments{
\item{object}{an \linkS4class{argoFloats} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Show some key facts about an \linkS4class{argoFloats} object.
}
\author{
Dan Kelley
}
