% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{useAdjustedNEW}
\alias{useAdjustedNEW}
\title{Switch [[ and Plot to Focus on Adjusted data}
\usage{
useAdjustedNEW(argos, fallback = "NA", debug = 0)
}
\arguments{
\item{argos}{an \code{\linkS4class{argoFloats}} object, as read by \code{\link[=readProfiles]{readProfiles()}}.}

\item{fallback}{a character value indicating what to do if all the adjusted
values for a particular parameter-profile pair are \code{NA}. The choices are
\code{"NA"} and \code{"raw"}; see \dQuote{Details}.}

\item{debug}{an integer that, if positive, indicates that some debugging information
should be printed.}
}
\description{
This function returns a copy of its first argument that has been modified by
(optionally) replacing 'raw' parameter values, and their flags, with
corresponding 'adjusted' values. The purpose is to ensure that future calls
to \code{\link{[[,argoFloats-method}} and plotting methods will focus on
the adjusted data instead of the raw data.  The procedure hinges on the value
of the \code{fallback} argument, as explained in \dQuote{Details}.
}
\details{
There are two cases, depending on the value of \code{fallback}.
\enumerate{
\item If \code{fallback} is \code{"NA"} (the default), then
the adjusted values become the focus, even if they are all \code{NA}. Thus,
\code{fallback="NA"} may be preferred as part of a cautious analysis that
focuses only on data of high quality. The downside of this approach is a
reduction in coverage, since it eliminates the 'raw' fields that are typical
of real-time mode datasets.
\item If \code{fallback} is \code{"raw"}, then the data in the adjusted parameter
data columns (corresponding to profiles) of each cycle are examined one by
one. If the values in a given column are all \code{NA}, then the parameter's raw
values for that column are not altered. However, if any of the values in a
given adjusted column are non-\code{NA}, then that entire 'adjusted' column is
copied into the corresponding 'raw' column. Since the \code{fallback="raw"}
scheme combines 'raw' and 'adjusted' data, it increases coverage, at the
possible expense of lowering overall data quality.
}
}
\author{
Dan Kelley and Jaimie Harbin
}
