% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,argoFloats-method}
\alias{show,argoFloats-method}
\title{Show Information About argoFloats Object}
\usage{
\S4method{show}{argoFloats}(object)
}
\arguments{
\item{object}{an \code{\linkS4class{argoFloats}} object.}
}
\value{
None (invisible NULL).
}
\description{
This produces a one-line explanation of the contents of the object,
that typically indicates the type and the number of items referenced
by the object. Those items depend on the type of object:
URLs if \code{metadata$type} is \code{"index"}, local filenames if
\code{"profiles"}, or oce-created argo objects if \code{"argos"}.
As with other R \code{show()} methods, it may be invoked in an interactive
session just by typing the object, or by using \code{print()}; see the Examples.
}
\examples{
library(argoFloats)
data(index)
show(index)
print(index)
index

}
\author{
Jaimie Harbin
}
