% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsGetFromCache}
\alias{argoFloatsGetFromCache}
\title{Get an item from the cache}
\usage{
argoFloatsGetFromCache(name, debug = 0)
}
\arguments{
\item{name}{character value, naming the item.}

\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\description{
Get an item from the cache
}
