% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{argoFloatsStoreInCache}
\alias{argoFloatsStoreInCache}
\title{Store an Item in the Cache}
\usage{
argoFloatsStoreInCache(name, value, debug = 0)
}
\arguments{
\item{name}{character value, naming the item.}

\item{value}{the new contents of the item.}

\item{debug}{an integer, passed to \code{\link[=argoFloatsDebug]{argoFloatsDebug()}}.}
}
\value{
None (invisible NULL).
}
\description{
Store an Item in the Cache
}
\seealso{
Other functions relating to cached values: 
\code{\link{argoFloatsClearCache}()},
\code{\link{argoFloatsGetFromCache}()},
\code{\link{argoFloatsIsCached}()},
\code{\link{argoFloatsListCache}()},
\code{\link{argoFloatsWhenCached}()}
}
\concept{functions relating to cached values}
