% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{D4900785_048.nc}
\alias{D4900785_048.nc}
\title{Sample Argo File (Delayed Core Data)}
\description{
This is NetCDF file for delayed-mode data for cycle 48 of Argo float 4900785, downloaded from
\code{https://data-argo.ifremer.fr/dac/aoml/4900785/profiles/D4900785_048.nc}
on 2020 June 24.
}
\examples{
library(argoFloats)
a <- readProfiles(system.file("extdata", "D4900785_048.nc", package = "argoFloats"))
summary(a)
summary(a[[1]])

}
\seealso{
Other raw datasets: 
\code{\link{R3901602_163.nc}},
\code{\link{SD5903586_001.nc}},
\code{\link{SR2902204_131.nc}}
}
\concept{raw datasets}
