% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getIndex}
\alias{getIndex}
\title{Get an Index of Available Argo Float Profiles}
\usage{
getIndex(
  server = "auto",
  filename = "argo",
  destdir = "~/data/argo",
  age = 7,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{server}{character value, or vector of character values, indicating
the name of servers that supply argo data.  If more than
one value is given, then these are tried sequentially until one
is found to supply the index file named in the \code{filename} argument.
As of May 2020, the two servers known to work are
\code{"ftp://usgodae.org/pub/outgoing/argo"} and
\code{"ftp://ftp.ifremer.fr/ifremer/argo"}.  These may be referred
to with nicknames \code{"usgodae"} and \code{"ifremer"}.  As a further
convenience, a third nickname (and the default for this argument)
is also available: \code{"auto"} is expanded to \code{c("usgodae","ifremer")}.
Note that if a nickname is not used, the character value(s) in \code{server}
must start with \code{"ftp://"}.}

\item{filename}{character value that indicates the file name on the server, as in
the first column of the table given in \dQuote{Details}, or (for some file types)
as in the nickname given in the middle column. Note that the downloaded
file name will be based on the full file name given as this argument, and
that nicknames are expanded to the full filenames before saving.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{age}{numeric value indicating how old a downloaded file
must be (in days), for it to be considered out-of-date.  The
default, \code{age=7}, limits downloads to once per week, as a way
to avoid slowing down a workflow with a download that might take
a sizable fraction of an hour. Set \code{age=0} to force a download,
regardless of the file age.}

\item{quiet}{silence some progress indicators.  The default
is to show such indicators.}

\item{debug}{integer value indicating level of debugging. If this
is less than 1, no debugging is done. Otherwise, some functions
will print debugging information.  If a function call fails, the
first step should be to rerun the function with \code{debug=1},
to see if the output suggests a problem in the call.}
}
\value{
An object of class \code{\linkS4class{argoFloats}} with type=\code{"index"}, which
is suitable as the first argument of \code{\link[=getProfiles]{getProfiles()}}.
}
\description{
This function gets an index of available Argo float profiles, either
by downloading information from a data repository or by reusing an index
(stored as an \code{.rda} file) that was prepared by a recent call to the function.
}
\details{
The first step is to construct a URL for downloading, based on the
\code{url} and \code{file} arguments. That URL will be a string ending in \code{.gz},
or \code{.txt} and from this the name of a local file is constructed by changing the
suffix to \code{.rda} and prepending the file directory specified by
\code{destdir}.  If an \code{.rda} file of that name already exists, and is less
than \code{age} days old, then no downloading takes place. This caching
procedure is a way to save time, because the download can take from a
minute to an hour, depending on the bandwidth of the connection to the
server.

The resultant \code{.rda} file, which is named in the return value of this
function, holds a list named \code{index} that holds following elements:
\itemize{
\item \code{ftpRoot}, the FTP root stored in the header of the source \code{file}
(see next paragraph).
\item \code{server}, the argument provided here.
\item \code{filename}, the argument provided here.
\item \code{header}, the preliminary lines in the source file that start
with the \verb{#} character.
\item \code{data}, a data frame containing the items in the source file.
The names of these items are determined automatically from
\code{"argo"} and \code{"bgcargo"} files, but for \code{"merged"} files,
the header is malformed (as of February 2020) and so the names
are set based on the authors' inspection of a downloaded file.
}

Some expertise is required in deciding on the value for the
\code{file} argument to \code{\link[=getIndex]{getIndex()}}.  As of February 2020, the
FTP site \verb{ftp://usgodae.org/pub/outgoing/argo} contains multiple
files that appear to be indices.  These are listed as the left-hand
column in the following table. The middle column lists nicknames
for the some files, which can be provided as the \code{file} argument,
as alternatives to the full names.
The right-hand column is a brief
description of the file contents, inferred by examination
of the file.  Note that there some files on the server
have names similar to those given below, but ending in \code{.txt} instead
of \code{.txt.gz}, but these files take longer to download and
seem to be equivalent to the \code{.gz} versions, so \code{\link[=getIndex]{getIndex()}} is
designed to work with them.  Also, note that, as of April 2020,
the usgodaee server does not supply \code{"synthetic"} files, but
the ifremer server does.
\tabular{lll}{
\emph{File Name}                           \tab \emph{Nickname}              \tab \emph{Contents}\cr
\code{ar_greylist.txt}                     \tab -                       \tab Suspicious/malfunctioning floats\cr
\code{ar_index_global_meta.txt.gz}         \tab -                       \tab Metadata files\cr
\code{ar_index_global_prof.txt.gz}         \tab \code{"argo"}                \tab Argo data\cr
\code{ar_index_global_tech.txt.gz}         \tab -                       \tab Technical files\cr
\code{ar_index_global_traj.txt.gz}         \tab -                       \tab Trajectory files\cr
\code{argo_bio-profile_index.txt.gz}       \tab \code{"bgc"} or \code{"bgcargo"}  \tab Biogeochemical Argo data (without S or T)\cr
\code{argo_bio-traj_index.txt.gz}          \tab -                       \tab Bio-trajectory files\cr
\code{argo_merge-profile_index.txt.gz}     \tab \code{"merge"} or \code{"merged"} \tab Merged \code{"argo"} and \code{"bgc"} data\cr
\code{argo_synthetic-profile_index.txt.gz} \tab \code{"synthetic"}           \tab Synthetic data, successor to \code{"merge"}\cr
}

The next step after using \code{\link[=getIndex]{getIndex()}} is usually to
use \code{\link[=getProfiles]{getProfiles()}}, which downloads or checks for local
copies of the per-profile data files that are listed in an
index, and this is typically followed by a call to
\code{\link[=readProfiles]{readProfiles()}}, which reads the local files, yielding
an object that can be plotted or analysed in other ways.
}
\examples{
\dontrun{
# Download an index of merged argo/bgc-argo floats, and plot temporal coverage.
library(argoFloats)
i <- getIndex(filename="merged")
summary(i)
hist(i[["date"]], breaks="years", main="", xlab="Time", freq=TRUE)}

}
\author{
Dan Kelley
}
