% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getIndex}
\alias{getIndex}
\title{Get an Index of Available Argo Float Profiles}
\usage{
getIndex(
  server = "auto",
  file = "argo",
  destdir = ".",
  age = 7,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{server}{character value, or vector of character values, giving
the base name(s) of server(s) holding argo profile files.  These servers
are tried sequentially until one of them works.  The default
value of \code{server} is \code{"auto"}, which is automatically
expanded to
\code{c("ftp://usgodae.org/pub/outgoing/argo", "ftp://ftp.ifremer.fr/ifremer/argo")},
meaning to try the USGODAE server first, but to switch to the IFREMER
server if that fails.}

\item{file}{character value that indicates the file name on the server, as in
the first column of the table given in \dQuote{Details}, or (for some file types)
as in the nickname given in the middle column. Note that the downloaded
file name will be based on the full file name given as this argument, and
that nicknames are expanded to the full filenames before saving.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{age}{numeric value indicating how old a downloaded file
must be (in days), for it to be considered out-of-date.  The
default, \code{age=7}, limits downloads to once per week, as a way
to avoid slowing down a workflow with a download that might take
a sizable fraction of an hour. Set \code{age=0} to force a download,
regardless of the file age.}

\item{quiet}{silence some progress indicators.  The default
is to show such indicators.}

\item{debug}{integer value indicating level of debugging. If this
is less than 1, no debugging is done. Otherwise, some functions
will print debugging information.  If a function call fails, the
first step should be to rerun the function with \code{debug=1},
to see if the output suggests a problem in the call.}
}
\value{
An object of class \linkS4class{argoFloats} with type=\code{"index"}, which
is suitable as the first argument of \code{\link[=getProfiles]{getProfiles()}}.
}
\description{
This function gets an index of available Argo float profiles, either by
by downloading information from a data repository or by reusing an index
(stored as an \code{.rda} file) that was prepared by a recent call to the function.
}
\details{
The first step is to construct a URL for downloading, based on the
\code{url} and \code{file} arguments. That URL will be a string ending in \code{.gz},
or \code{.txt} and from this the name of a local file is constructed by changing the
suffix to \code{.rda} and prepending the file directory specified by
\code{destdir}.  If an \code{.rda} file of that name already exists, and is less
than \code{age} days old, then no downloading takes place. This caching
procedure is a way to save time, because the download can take from a
minute to an hour, depending on the bandwidth of the connection to the
server.

The resultant \code{.rda} file, which is named in the return value of this
function, holds a list named \code{index} that holds following elements:
\itemize{
\item \code{ftpRoot}, the FTP root stored in the header of the source \code{file}
(see next paragraph).
\item \code{server}, the argument provided here.
\item \code{file}, the argument provided here.
\item \code{header}, the preliminary lines in the source file that start
with the \verb{#} character.
\item \code{data}, a data frame containing the items in the source file.
The names of these items are determined automatically from
\code{"argo"} and \code{"bgcargo"} files, but for \code{"merged"} files,
the header is malformed (as of February 2020) and so the names
are set based on the authors' inspection of a downloaded file.
}

Note that \code{paste0(argoIndex$ftpRoot, argoIndex$data$file)} will
form a vector of URLs pointing to argo netcdf files that can be
downloaded with any system tool for downloading files, but it
is more practical to use \code{\link[=getProfileFromUrl]{getProfileFromUrl()}}, which also checks
to see if the data has already been constructed for a given URL.
Even better, one should use \code{\link[=getProfiles]{getProfiles()}}, which computes
a set of relevant URLs based on the contents of the index, and
places the results in the \code{destfile} directory that was supplied
to \code{\link[=getIndex]{getIndex()}}, and stored in the index.

Some expertise is required in deciding on the value for the
\code{file} argument to \code{\link[=getIndex]{getIndex()}}.  As of February 2020, the
FTP site \verb{ftp://usgodae.org/pub/outgoing/argo} contains multiple
files that appear to be indices.  These are listed as the left-hand
column in the following table. The middle column is a "nickname"
for the file (and can be provided as the \code{file} argument to the
\code{\link[=getIndex]{getIndex()}} function).  The right-hand column is a brief
description of the file contents, inferred by examination
of the file.  (Note that there some files on the server
have names similar to those given below, but ending in \code{.txt} instead
of \code{.txt.gz}, but these files take longer to download and
seem to be equivalent to the \code{.gz} versions, so \code{\link[=getIndex]{getIndex()}} is
designed to work with them.)
\tabular{lll}{
\emph{File Name}                       \tab \emph{Nickname}              \tab \emph{Contents}\cr
\code{ar_greylist.txt}                 \tab -                       \tab Suspious or malfunctioning float sensors.\cr
\code{ar_index_global_meta.txt.gz}     \tab -                       \tab Metadata files of the argo GDAC ftp site.\cr
\code{ar_index_global_prof.txt.gz}     \tab \code{"argo"}                \tab Argo data.\cr
\code{ar_index_global_tech.txt.gz}     \tab -                       \tab All technical files of the argo GDAC ftp site.\cr
\code{ar_index_global_traj.txt.gz}     \tab -                       \tab All trajectory files of the argo GDAC ftp site.\cr
\code{argo_bio-profile_index.txt.gz}   \tab \code{"bgc"} or \code{"bgcargo"}  \tab Biogeochemical Argo data (without S or T).\cr
\code{argo_bio-traj_index.txt.gz}      \tab -                       \tab Bio-trajectory files of the argo GDAC ftp site.\cr
\code{argo_merge-profile_index.txt.gz} \tab \code{"merge"} or \code{"merged"} \tab Merged \code{"argo"} and \code{"bgc"} data.\cr
}
}
\examples{
\dontrun{
library(argoFloats)
library(oce)
# Example: Temporal and spatial coverage of merged argo/bgcargo measurements.
par(mfrow=c(2, 1), mar=c(3, 3, 1, 1))
ai <- getIndex(file="merged", destdir="~/data/argo")
summary(ai)
hist(ai[["date"]], breaks="years", main="", xlab="Time", freq=TRUE)
data(coastlineWorld)
par(mar=rep(0.5, 4))
mapPlot(coastlineWorld, col="gray")
mapPoints(ai[["longitude"]], ai[["latitude"]], pch=".", col="blue")
}

}
\author{
Dan Kelley
}
