% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getProfileFromUrl}
\alias{getProfileFromUrl}
\title{Get Data for an Argo Float Profile}
\usage{
getProfileFromUrl(
  url = NULL,
  destdir = "~/data/argo",
  destfile = NULL,
  force = FALSE,
  retries = 3,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{url}{character value giving the URL for an Argo float a profile.}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{destfile}{optional character value that specifies the name to be used
for the downloaded file. If this is not specified, then a name is determined
from the value of \code{url}.}

\item{force}{a logical value that indicates whether to force a file
download, even if the destination file already exists.}

\item{retries}{integer telling how many times to retry a download,
if the first attempt fails.}

\item{quiet}{logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}},
that indicates whether to download silently.  Since some downloads
can be slow, the default value is \code{quiet=FALSE}.}

\item{debug}{integer value indicating level of debugging. If this
is less than 1, no debugging is done. Otherwise, some functions
will print debugging information.  If a function call fails, the
first step should be to rerun the function with \code{debug=1},
to see if the output suggests a problem in the call.}
}
\description{
Get Data for an Argo Float Profile
}
\examples{
\dontrun{
# These examples assume that the ~/data/argo directory exists.
library(argoFloats)
library(oce)

# Example 1: a particular file
url <- "ftp://ftp.ifremer.fr/ifremer/argo/dac/nmdis/2901633/profiles/R2901633_071.nc"
file <- getProfileFromUrl(url=url, destdir="~/data/argo")
argo <- read.argo(file)
plot(argo, which=c(1, 4, 6, 5))

# Example 2: argo profile nearest Sable Island
index <- getIndex()
lon0 <- -59.9149
lat0 <- 43.9337
if (requireNamespace("oce")) {
    dist <- oce::geodDist(index[["longitude"]], index[["latitude"]], lon0, lat0)
    w <- which.min(dist)
    url <- paste0(index[["metadata"]]["ftpRoot"], "/", index[["file"]][w])
    fileSable <- getProfileFromUrl(url=url)
    argoSable <- read.oce(fileSable)
    plot(argoSable, which=c(1, 4, 6, 5))
}
}

}
\author{
Dan Kelley
}
