% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getProfiles}
\alias{getProfiles}
\title{Get Data for an Argo Float Profile}
\usage{
getProfiles(
  url = NULL,
  destdir = ".",
  destfile,
  mode = "wb",
  retries = 3,
  force = FALSE,
  quiet = FALSE,
  debug = 0
)
}
\arguments{
\item{url}{character value giving the URLs for an Argo float a profile. (In
a future version, \code{url} may be permitted to be a vector.)}

\item{destdir}{character value indicating the directory in which to store
downloaded files. The default value of \code{"."} means to store the
downloaded file in the present working directory.  Set \code{destdir=NULL}
if \code{destfile} is a filename with full path information.
File clutter is reduced by creating a top-level directory called
\code{data}, with subdirectories for various file types; see
\dQuote{Examples}.}

\item{destfile}{optional character value that specifies the name to be used
for the downloaded file. If this is not specified, then a name is determined
from the value of \code{url}.}

\item{mode}{character value passed to \code{\link[curl:curl_download]{curl::curl_download()}},
controlling the mode used in writing the local file. The default
value, \code{mode="wb"}, means to set the write format to binary. This
is necessary on Windows machines, which otherwise will alter newline
characters, ruining file integrity for binary files (such as
netcdf files). The setting is not required on other machines, but
it makes sense to use this default in all instances.  See the
documentation for \code{\link[curl:curl_download]{curl::curl_download()}}.}

\item{retries}{integer telling how many times to retry a download,
if the first attempt fails.}

\item{force}{A logical value that indicates whether to force the download,
even if the pathname constructed from \code{destdir} and \code{destfile}
already exists.}

\item{quiet}{logical value, passed to \code{\link[curl:curl_download]{curl::curl_download()}},
that indicates whether to download silently.  Since some downloads
can be slow, the default value is \code{quiet=FALSE}.}

\item{debug}{integer value indicating level of debuggin. If this
is less than 1, no debugging is done. Otherwise, some functions
will print debugging information.  If a function call fails, the
first step should be to rerun the function with \code{debug=1},
to see if the output suggests a problem in the call.}
}
\description{
Get Data for an Argo Float Profile
}
\examples{
\dontrun{
# These examples assume that the ~/data/argo directory exists and is readable.
library(argoFloats)
library(oce)
#
# Example 1: get float profile based on URL
url <- "ftp://ftp.ifremer.fr/ifremer/argo/dac/nmdis/2901633/profiles/R2901633_071.nc"
file <- getProfiles(url=url, destdir="~/data/argo")
argo <- read.argo(file)
plot(argo, which=c(1, 4, 6, 5))
#
# Example 2: get float profile nearest Sable Island
theIndex <- getIndex(destdir="~/data/argo")
load(theIndex) # defines 'index'
lon0 <- -59.9149
lat0 <- 43.9337
dist <- geodDist(index$data$longitude, index$data$latitude, lon0, lat0)
which0 <- which.min(dist)
url <- paste0(index$ftpRoot, "/", index$data$file[which0])
file0 <- getProfiles(url=url, destdir="~/data/argo")
argo0 <- read.oce(file0)
plot(argo0, which=c(1, 4, 6, 5))
}
}
\author{
Dan Kelley
}
