% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hexToNibble}
\alias{hexToNibble}
\title{Convert a hex digit to integer vector of length 4}
\usage{
hexToNibble(x)
}
\arguments{
\item{x}{A character value corresponding to a hex digit (i.e. 0 through 9, or A through F).}
}
\value{
An integer vector.
}
\description{
This function converts hex digits to 4 integers indicating bits, in 'math' order. It's
intended to be used to look at the hex digits of \code{HISTORY_QCTEST} of an \code{\linkS4class{argoFloats}}
object that was created by \code{\link[=readProfiles]{readProfiles()}}. It then converts these hex digits to bits to
be used within the \code{\link[=showQCTests]{showQCTests()}} function.
}
\examples{
library(argoFloats)
hexToNibble('2')
hexToNibble('e')

}
\author{
Jaimie Harbin and Dan Kelley
}
