% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hexToNibble}
\alias{hexToNibble}
\title{Convert a hex digit to 4 integers indicating the bits, in 'math' order.}
\usage{
hexToNibble(x)
}
\arguments{
\item{x}{a character value corresponding to a hex digit (i.e. 0 through 9, or A through F)}
}
\value{
an integer vector
}
\description{
This function converts hex digits to 4 integers indicating bits, in 'math' order. Its
intended to be used to look at the hex digits of \code{HISTORY_QCTEST} from an argos object
that was created by \code{\link[=readProfiles]{readProfiles()}}. It then converts these hex digits to bits to
be used in the \code{\link[=showQCTests]{showQCTests()}} function.
}
\examples{
hexToNibble('2') # 0 0 1 0
hexToNibble('e') # 1 1 1 0
}
