% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hexToNibble}
\alias{hexToNibble}
\title{Convert Hexadecimal Digit to Integer Vector}
\usage{
hexToNibble(x)
}
\arguments{
\item{x}{A character value corresponding to a hexadecimal digit (i.e. \code{"0"} through \code{"9"},
\code{"a"} through \code{"f"}, or \code{"A"} through \code{"F"}).}
}
\value{
An integer vector, in 'mathematical' order.  (This is the reverse of
the order used by \code{\link[=rawToBits]{rawToBits()}}; see the \dQuote{Examples}.)
}
\description{
\code{hexToNibble} converts a hexadecimal digit to 4 integers indicating bits, e.g. for use within
\code{\link[=showQCTests]{showQCTests()}}.
}
\examples{
library(argoFloats)
hexToNibble('1') # 0 0 0 1
hexToNibble('e') # 1 1 1 0

}
\author{
Jaimie Harbin and Dan Kelley
}
