% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{data}
\name{index}
\alias{index}
\title{A sample index of profiles}
\description{
This is created by subsetting a global index to the 1788 Argo profiles
that were within a 200km radius of Marsh Harbour, Abaco Island,
Bahamas, as of 2020 March 14, using
the following code.
\preformatted{
library(oce)
indexAll <- getIndex()
index <- subset(indexAll,
    circle=list(longitude=-77.06,latitude=26.54,radius=200))
}
}
\examples{
library(argoFloats)
data(index, package="argoFloats")
plot(index, which="map")

}
\seealso{
Other datasets provided with argoFloats: 
\code{\link{indexBgc}},
\code{\link{indexMerged}}
}
\concept{datasets provided with argoFloats}
