% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{data}
\name{indexBgc}
\alias{indexBgc}
\title{A sample index of biogeochemical-argo profiles}
\description{
This is created by subsetting a global index to the 39 BGC Argo profiles
that were within a 300km radius of Marsh Harbour, Abaco Island,
Bahamas, as of 2020 March 23, using
the following code.
\preformatted{
library(argoFloats)
indexAll <- getIndex(file="bgc")
indexBgc <- subset(indexAll, circle=list(longitude=-77.06, latitude=26.54, radius=300))
save(indexBgc, file="indexBgc.rda")
tools::resaveRdaFiles('indexBgc.rda')
}
}
\examples{
library(argoFloats)
data(indexBgc, package="argoFloats")
plot(indexBgc, which="map")
# Tabulate BGC parameters in these files
table(indexBgc[["parameters"]])

}
\seealso{
Other datasets provided with argoFloats: 
\code{\link{indexMerged}},
\code{\link{index}}
}
\concept{datasets provided with argoFloats}
