% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{data}
\name{indexMerged}
\alias{indexMerged}
\title{A sample index of merged argo and biogeochemical-argo profiles}
\description{
This was created by subsetting a global index to the 39 BGC Argo profiles
that were within a 300km radius of Marsh Harbour, Abaco Island,
Bahamas, as of 2020 March 23, using the following code.\preformatted{library(argoFloats)
indexAll <- getIndex("merged")
indexMerged <- subset(indexAll,
    circle=list(longitude=-77.06, latitude=26.54, radius=300))
}
}
\section{Historical note}{

This "merged" file from the usgodae server is likely to be removed, when
that server changes to the "synthetic" file format that the ifremer server
uses (as of May, 2020 and perhaps months previously, since the changeover
data was supposed to be Dec, 2019).
}

\examples{
library(argoFloats)
data(indexMerged)
plot(indexMerged)
summary(indexMerged)
unique(indexMerged[["parameters"]])
}
\seealso{
Other datasets provided with argoFloats: 
\code{\link{indexBgc}},
\code{\link{indexSynthetic}},
\code{\link{index}}
}
\concept{datasets provided with argoFloats}
