% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,argoFloats,ANY-method}
\alias{plot,argoFloats,ANY-method}
\alias{plot,argoFloat-method}
\title{Plot an argoFloats object}
\usage{
\S4method{plot}{argoFloats,ANY}(
  x,
  which = "map",
  cex = NULL,
  col = NULL,
  pch = NULL,
  mar = NULL,
  mgp = NULL,
  debug = 0,
  ...
)
}
\arguments{
\item{x}{an \linkS4class{argoFloats} object.}

\item{which}{a string that indicates the type of plot; see \dQuote{Details}.}

\item{cex}{character expansion factor for plot symbols, or \code{NULL}, to get an
value that depends on the value of \code{which}.}

\item{col}{colour to be used for plot symbols, or \code{NULL}, to get an
value that depends on the value of \code{which}.}

\item{pch}{number indicating the type of plot symbol, or \code{NULL}, to get an
value that depends on the value of \code{which}.}

\item{mar}{either a four-element vector giving the \code{par("mar")} values to
be used for the plot, or \code{NULL}, which means that this function will determine
the margin based on the plot type.}

\item{mgp}{similar to \code{mar}, but for \code{par("mgp")}.}

\item{debug}{an integer specifying the level of debugging.}

\item{\dots}{extra arguments passed to the plot calls that are made
to within this function. A common use for \code{which="map"} cases is
to set \code{xlim} and \code{ylim} to get enlarge the plot region, so a coastline
becomes visible.}
}
\description{
The action depends on the \code{type} of the object, and
this is set up by the function that created the object;
see \dQuote{Details}.
}
\details{
The various plot types are as follows.
\itemize{
\item for \code{which="map"}, a map of profile locations is created. This
only works if the \code{type} is \code{"index"} (meaning that \code{x} was created
by \code{\link[=getIndex]{getIndex()}} or a subset of such an object, created with
\code{\link[=subset,argoFloats-method]{subset,argoFloats-method()}}), or \code{argos} (meaning that
\code{x} was created with \code{\link[=readProfiles]{readProfiles()}}.  The plot range is
auto-selected.  If the \code{ocedata} package
is available, then its \code{coastlineWorldFine} dataset is used to draw
a coastline (which will be visible only if the plot region
is large enough); otherwise, if the \code{oce} package is available, then its
\code{coastlineWorld} dataset is used.
}
}
\examples{
# Example 1
library(argoFloats)
data(index)
plot(index, which="map")

}
\author{
Dan Kelley
}
