% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readProfiles}
\alias{readProfiles}
\title{Read argo profiles from local files}
\usage{
readProfiles(profiles, debug = 0)
}
\arguments{
\item{profiles}{either a character vector holding the names
of local files to read, or (better) an \linkS4class{argoFloats} object created
by \code{\link[=getProfiles]{getProfiles()}}.}

\item{debug}{an integer specifying the level of debugging. If
this is zero, the work proceeds silently. If it is 1,
a small amount of debugging information is printed.  Note that
\code{debug-1} is passed to \code{\link[oce:read.argo]{oce::read.argo()}}, which actually reads
the file, and so it will print messages if \code{debug} exceeds 1.}
}
\value{
\link{readProfiles} returns an \linkS4class{argoFloats} object
with \code{type="argo"}, in which the \code{data} slot
contains a list named \code{argos} that holds objects
that are created by \code{\link[oce:read.argo]{oce::read.argo()}}.
}
\description{
This works with either a list of local (netCDF) files,
or a \linkS4class{argoFloats} object of type \code{"profiles"}, as
created by \code{\link[=getProfiles]{getProfiles()}}.
}
\examples{
\dontrun{
library(argoFloats)
library(oce)
# TS diagram using Argo profiles near Sable Island.
indexAll <- getIndex(destdir="~/data/argo")
# The next line yields 63 profiles as of February 2020
index <- subset(indexAll, circle=list(longitude=-59.915, latitude=44.934, radius=180))
profiles <- getProfiles(index)
argos <- readProfiles(profiles)
# Extract and plot
for (i in seq_len(argos[["profile count"]])) {
    argo <- argos[["profile", i]]
    if (i == 1) {
        plotTS(argo, Slim=c(31, 36), Tlim=c(1,24), eos="gsw")
    } else {
        SA <- argo[["SA"]]
        CT <- argo[["CT"]]
        points(SA, CT, col=i\%\%10)
    }
}
}

}
\author{
Dan Kelley
}
