% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readProfiles}
\alias{readProfiles}
\title{Read argo profiles from local files}
\usage{
readProfiles(
  profiles,
  handleFlags,
  adjusted = TRUE,
  FUN,
  silent = FALSE,
  debug = 0
)
}
\arguments{
\item{profiles}{either a character vector holding the names
of local files to read, or (better) an \code{\linkS4class{argoFloats}} object created
by \code{\link[=getProfiles]{getProfiles()}}.}

\item{handleFlags}{an optional logical value that is set to \code{FALSE}
if not provided (with a message being indicated to that effect).
See \dQuote{Details}.}

\item{adjusted}{a logical value (\code{TRUE} by default) that indicates
whether to focus on the "adjusted" versions of data and quality-control
flags, renaming the other versions with names ending in \code{Unadjusted}.
See \dQuote{Details}.}

\item{FUN}{a function that reads the netcdf files in which the argo
profiles are stored.  If \code{FUN} not provided, then it defaults
to \code{\link[oce:read.argo]{oce::read.argo()}}.  Only experts should consider anything
other than this default, or a wrapper to it.}

\item{silent}{logical value (with \code{FALSE} as the default value),
indicating whether to silence messages and warnings. The value of
\code{silent} is ignored if \code{debug} is \code{TRUE}.}

\item{debug}{an integer specifying the level of debugging. If
this is zero, the work proceeds silently. If it is 1,
a small amount of debugging information is printed.  Note that
\code{debug-1} is passed to \code{\link[oce:read.argo]{oce::read.argo()}}, which actually reads
the file, and so it will print messages if \code{debug} exceeds 1.}
}
\value{
\link{readProfiles} returns an \code{\linkS4class{argoFloats}} object
with \code{type="argos"}, in which the \code{data} slot
contains a list named \code{argos} that holds objects
that are created by \code{\link[oce:read.argo]{oce::read.argo()}}.
}
\description{
This works with either a list of local (netCDF) files,
or a \code{\linkS4class{argoFloats}} object of type \code{"profiles"}, as
created by \code{\link[=getProfiles]{getProfiles()}}.  By default, warnings are issued about any
profiles in which 10 percent or more of the measurements are flagged
with a quality-control code of 4 (which designates bad data),
and these values are set to \code{NA}; use the \code{silent} and \code{handleFlags}
arguments to control this behaviour.
}
\details{
If \code{handleFlags} is \code{TRUE}, then the \code{\link[oce:handleFlags]{oce::handleFlags()}} function
from the \CRANpkg{oce} package is called on each individual argo object
that is read by \code{\link[oce:read.argo]{oce::read.argo()}} in that package.  The action
of \code{\link[oce:handleFlags]{oce::handleFlags()}} is see whether any data are flagged
with the quality-control code that is not equal to 1, and to set
the corresponding data to \code{NA}.  In the \code{oce} terminology,
the flag meanings are:
\code{not_assessed}=0, \code{passed_all_tests}=1, \code{probably_good}=2, \code{probably_bad}=3,
\code{bad}=4, \code{changed}=5, \code{averaged}=7, \code{interpolated}=8, \code{missing}=9, and so
the restriction to only values equal to 1 means that multiple categories
of potentially useable data are discarded, and for that reason,
\code{handleFlags} is set to \code{FALSE} unless the user provides a value in
the call to \code{readProfiles()}.  See Wong et al. (2020) for a discussion
of flags in argo data.

If \code{adjusted} is \code{TRUE}, then the data elements are renamed after
reading, so that e.g. the data named \code{TEMP_ADJUSTED} and
\code{TEMP_ADJUSTED_ERROR} in the source netcdf file would be
stored with names \code{temperature} and \code{temperatureAdjusted},
with \code{TEMP} being stored as \code{temperatureUnadjusted}.  This
applies to all variables, not just temperature.  The quality-control
flags are similarly named. Note that the original names are left
intact, so that with\preformatted{a <- readProfiles(..., adjusted=TRUE)
}

the values of \code{a[["temperature"]]} and \code{a[["TEMP_ADJUSTED"]]}
will be identical.
}
\examples{
# Download and plot some profiles.
\dontrun{
library(argoFloats)
data(index)
index1 <- subset(index, 1)
profiles <- getProfiles(index1)
argosWithNA<- readProfiles(profiles, handleFlags=FALSE)
argosWithoutNA <- readProfiles(profiles, handleFlags=FALSE)
par(mfrow=c(1, 2))
file <- gsub(".*/", "",  profiles[[1]])
aWithNA <- argosWithNA[[1]]
plotTS(aWithNA, eos="unesco", type="o")
mtext(paste(file, "\n handling flags"), cex=0.7*par("cex"))
aWithoutNA <- argosWithoutNA[[1]]
plotTS(aWithoutNA, eos="unesco", type="o")
mtext(paste(file, "\n ignoring flags"), cex=0.7*par("cex"))
}

}
\references{
Wong, Annie, Robert Keeley, Thierry Carval, and Argo Data Management Team.
“Argo Quality Control Manual for CTD and Trajectory Data,” January 1, 2020.
\url{https://archimer.ifremer.fr/doc/00228/33951}.
\url{http://dx.doi.org/10.13155/33951}.
}
\author{
Dan Kelley
}
