% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readProfiles}
\alias{readProfiles}
\title{Read argo profiles from local files}
\usage{
readProfiles(profiles, debug = 0)
}
\arguments{
\item{profiles}{either a character vector holding the names
of local files to read, or (better) an \linkS4class{argoFloats} object created
by \code{\link[=getProfiles]{getProfiles()}}.}

\item{debug}{an integer specifying the level of debugging. If
this is zero, the work proceeds silently. If it is 1,
a small amount of debugging information is printed.  Note that
\code{debug-1} is passed to \code{\link[oce:read.argo]{oce::read.argo()}}, which actually reads
the file, and so it will print messages if \code{debug} exceeds 1.}
}
\value{
\link{readProfiles} returns an \linkS4class{argoFloats} object
with \code{type="argo"}, in which the \code{data} slot
contains a list named \code{argos} that holds objects
that are created by \code{\link[oce:read.argo]{oce::read.argo()}}.
}
\description{
This works with either a list of local (netCDF) files,
or a \linkS4class{argoFloats} object of type \code{"profiles"}, as
created by \code{\link[=getProfiles]{getProfiles()}}.
}
\examples{
# Download and plot some profiles.
\donttest{
library(argoFloats)
data(index)
index4 <- subset(index, 1:4)
profiles <- getProfiles(index4, destdir=".")
argos <- readProfiles(profiles)
par(mfrow=c(2, 2))
library(oce)
for (i in 1:2) {
  A <- argos[[i]]
  filename <- gsub(".*/", "", A[["filename"]])
  plotTS(A, eos="unesco")
  mtext(filename, line=0.7, cex=par("cex"))
  Aclean <- handleFlags(A)
  plotTS(Aclean, eos="unesco")
  mtext(paste(filename, "(cleaned)"), line=0.7, cex=par("cex"))
}
}

}
\author{
Dan Kelley
}
