% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readProfiles}
\alias{readProfiles}
\title{Read argo profiles from local files}
\usage{
readProfiles(profiles, handleFlags = TRUE, debug = 0)
}
\arguments{
\item{profiles}{either a character vector holding the names
of local files to read, or (better) an \code{\linkS4class{argoFloats}} object created
by \code{\link[=getProfiles]{getProfiles()}}.}

\item{handleFlags}{a logical value (with default \code{TRUE}) that
indicates whether to call \code{\link[oce:handleFlags]{oce::handleFlags()}} on the individual argo
files that are to be read. This cleans up some common errors
that are identified in the quality-control analysis that is usually
done before data are place on the argo servers.}

\item{debug}{an integer specifying the level of debugging. If
this is zero, the work proceeds silently. If it is 1,
a small amount of debugging information is printed.  Note that
\code{debug-1} is passed to \code{\link[oce:read.argo]{oce::read.argo()}}, which actually reads
the file, and so it will print messages if \code{debug} exceeds 1.}
}
\value{
\link{readProfiles} returns an \code{\linkS4class{argoFloats}} object
with \code{type="argos"}, in which the \code{data} slot
contains a list named \code{argos} that holds objects
that are created by \code{\link[oce:read.argo]{oce::read.argo()}}.
}
\description{
This works with either a list of local (netCDF) files,
or a \code{\linkS4class{argoFloats}} object of type \code{"profiles"}, as
created by \code{\link[=getProfiles]{getProfiles()}}.
}
\examples{
# Download and plot some profiles.
\dontrun{
library(argoFloats)
data(index)
index2 <- subset(index, 1:2)
profiles <- getProfiles(index2)
argos <- readProfiles(profiles, handleFlags=TRUE)
par(mfrow=c(1, 2))
library(oce)
for (i in 1:2) {
  A <- argos[["profile", i]]
  filename <- gsub(".*/", "", A[["filename"]])
  plotTS(A, eos="unesco")
  mtext(filename, line=0.7, cex=par("cex"))
}
}

}
\author{
Dan Kelley
}
