% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readProfiles}
\alias{readProfiles}
\title{Read argo profiles from local files}
\usage{
readProfiles(profiles, adjusted = FALSE, FUN, silent = FALSE, debug = 0)
}
\arguments{
\item{profiles}{either a character vector holding the names
of local files to read, or (better) an \code{\linkS4class{argoFloats}} object created
by \code{\link[=getProfiles]{getProfiles()}}.}

\item{adjusted}{a logical value (\code{FALSE} by default) that indicates
whether to focus on the "adjusted" versions of data and quality-control
flags, renaming the other versions with names ending in \code{Unadjusted}.
See \dQuote{Details}.}

\item{FUN}{a function that reads the netcdf files in which the argo
profiles are stored.  If \code{FUN} not provided, then it defaults
to \code{\link[oce:read.argo]{oce::read.argo()}}.  Only experts should consider anything
other than this default, or a wrapper to it.}

\item{debug}{an integer specifying the level of debugging. If
this is zero, the work proceeds silently. If it is 1,
a small amount of debugging information is printed.  Note that
\code{debug-1} is passed to \code{\link[oce:read.argo]{oce::read.argo()}}, which actually reads
the file, and so it will print messages if \code{debug} exceeds 1.}
}
\value{
\link{readProfiles} returns an \code{\linkS4class{argoFloats}} object
with \code{type="argos"}, in which the \code{data} slot
contains a list named \code{argos} that holds objects
that are created by \code{\link[oce:read.argo]{oce::read.argo()}}.
}
\description{
This works with either a list of local (netCDF) files,
or a \code{\linkS4class{argoFloats}} object of type \code{"profiles"}, as
created by \code{\link[=getProfiles]{getProfiles()}}.
During the reading, argo profile objects are created with \code{\link[oce:read.argo]{oce::read.argo()}},
after which quality-control flags are inserted into the objects for later use
by \code{\link[=applyQC]{applyQC()}}, the documentation of which explains the meanings of the flags.
}
\details{
By default, warnings are issued about any
profiles in which 10 percent or more of the measurements are flagged
with a quality-control code of 0, 3, 4, 6, 7, or 9 (see the
\code{\link[=applyQC]{applyQC()}} documentation for the meanings of these codes).
}
\examples{
# Read 5 profiles and plot TS for the first, in raw and QC-cleaned forms.
\dontrun{
library(argoFloats)
data(index)
index1 <- subset(index, 1)
profiles <- getProfiles(index1)
raw <- readProfiles(profiles)
clean <- applyQC(profiles)
par(mfrow=c(1, 2))
file <- gsub(".*/", "",  profiles[[1]])
aWithNA <- argosWithNA[[1]]
plotTS(raw[[1]], eos="unesco", type="o")
mtext(file, cex=0.7*par("cex"))
aWithoutNA <- argosWithoutNA[[1]]
plotTS(clean[[1]], eos="unesco", type="o")
mtext(paste(file, "\n (after applying QC)"), cex=0.7*par("cex"))
}

}
\author{
Dan Kelley
}
