% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showqc.R
\name{showQCTests}
\alias{showQCTests}
\title{Show which QC tests were performed and failed on an argos class}
\usage{
showQCTests(a)
}
\arguments{
\item{a}{a single float of \code{\linkS4class{argoFloats}}}
}
\value{
The associated quality-control (QC) test numbers performed and failed.
}
\description{
This function uses integer value from \code{\link[=hexToNibble]{hexToNibble()}} internally to
convert hex digits of \code{HISTORY_QCTEST} of a single \code{\linkS4class{argoFloats}}
object that was created by \code{\link[=readProfiles]{readProfiles()}} to indicate which QC tests
were performed and/or failed.
}
\examples{
\dontrun{
library(argoFloats)
data('index')
subset <- subset(index, 1)
profiles <- getProfiles(subset)
argos <- readProfiles(profiles)
argos1 <- argos[[1]]
showQCTests(argos[[1]])
  } 
}
