% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\encoding{UTF-8}
\name{SR2902204_131.nc}
\alias{SR2902204_131.nc}
\title{Sample Argo File (Real-Time Synthetic Data)}
\description{
This is the NetCDF file for cycle 131 of Argo float 2902204, downloaded from
\code{ftp://ftp.ifremer.fr/ifremer/argo/dac/incois/2902204/profiles/SR2902204_131.nc}
on 2020 June 24.
As its filename indicates, it holds the "synthetic" version of "real-time" data.
}
\examples{
library(argoFloats)
a <- readProfiles(system.file("extdata", "SR2902204_131.nc", package = "argoFloats"))
summary(a)
summary(a[[1]])

}
\seealso{
Other raw datasets: 
\code{\link{D4900785_048.nc}},
\code{\link{R3901602_163.nc}},
\code{\link{SD5903586_001.nc}}
}
\concept{raw datasets}
