% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{[[,argoFloats,ANY,ANY-method}
\alias{[[,argoFloats,ANY,ANY-method}
\alias{[[,argoFloats-method}
\title{Look up a Value Within an argoFloats Object}
\usage{
\S4method{[[}{argoFloats,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{an \code{\linkS4class{argoFloats}} object.}

\item{i}{a character value that specifies the item to be looked up;
see \dQuote{Details}.}

\item{j}{supplemental index value, used for some \code{x} types
and \code{i} values; see \dQuote{Details}.}

\item{...}{ignored.}
}
\description{
This function provides an easy way to look up values within an \code{\linkS4class{argoFloats}}
object, without the need to know the exact structure of the data. The action
taken by \code{[[} depends on the \code{type} element in the \code{metadata} slot of
the object (which is set by the function that created the object), on the
value of \code{i} and, in some cases, on the value of \code{j}; see \dQuote{Details}.
}
\details{
There are several possibilities, depending on the object \code{type}, as
listed below. Note that these possibilities are checked in the
order listed here.
\enumerate{
\item For all object types:
\enumerate{
\item If \code{i} is \code{"metadata"} then the \code{metadata} slot of \code{x} is returned.
\item Otherwise, if \code{i} is \code{"data"} then the \code{data} slot of \code{x} is returned.
\item Otherwise, the following steps are taken, depending on \code{type}.
}
\item If \code{type} is \code{"index"}, i.e. if \code{x} was created with \code{\link[=getIndex]{getIndex()}}
or with \code{\link[=subset,argoFloats-method]{subset,argoFloats-method()}} acting on the result of \code{\link[=getIndex]{getIndex()}},
then:
\enumerate{
\item If \code{i} is numeric and \code{j} is unspecified, then \code{i} is taken to
be an index that identifies the row(s) of the data frame that
was constructed by \code{\link[=getIndex]{getIndex()}} based on the remote index file
downloaded from the Argo server.  This has elements \code{file} for
the remote file name, \code{date} for the date of the entry, \code{latitude}
and \code{longitude} for the float position, etc.
\item If \code{i} is the name of an item in the \code{metadata} slot, then that item
is returned. The choices are:
\code{"destdir"}, \code{"destfileRda"}, \code{"filename"}, \code{"ftpRoot"}, \code{"header"},
\code{"server"}, \code{"type"}, and \code{"url"}.
\item Otherwise, if \code{i} is the name of an item in the \code{data} slot, then that item
is returned.  The choices are:
\code{"date"}, \code{"date_update"}, \code{"file"}, \code{"institution"}, \code{"latitude"},
\code{"longitude"}, \code{"ocean"}, and \code{"profiler_type"}.
\item Otherwise, if \code{i=="index"} then that item from the \code{data} slot of \code{x} is returned.
(For the possible names, see the previous item in this sub-list.)
\item Otherwise, if \code{i} is an integer, then the \code{i}-th row of the \code{index} item in
the \code{data} slot is returned.  This is a good way to learn the
longitude and latitude of the profile, the file name on the server,
etc.
\item Otherwise, if \code{i} is \code{"ID"} then the return value is developed from the
\code{index$file} item within the \code{data} slot of \code{x}, in one of three cases:
\enumerate{
\item If \code{j} is not supplied, the return value is a vector holding the
identifiers (character codes for numbers) for all the data files
referred to in the index.
\item Otherwise, if \code{j} is numeric, then the return value is a subset of
the ID codes, as indexed by \code{j}.
\item Otherwise, an error is reported.
}
\item If \code{i} is \code{"length"}, the number of remote files pointed to by the index
is returned.
}
\item Otherwise, if \code{type} is \code{"profiles"}, i.e. if \code{x} was created with \code{\link[=getProfiles]{getProfiles()}}, then:
\enumerate{
\item If \code{i} is numeric and \code{j} is unspecified, then return the local file name(s)
that are identified by using \code{i} as an index.
\item If \code{i} is the name of an item in the \code{metadata} slot, then that item
is returned. The choices are:
\code{"type"} and \code{"destdir"}.
\item Otherwise, if \code{i} is the name of an item in the \code{data} slot, then that item
is returned.  There is only one choice: \code{"file"}.
\item Otherwise, if \code{i} is \code{"profile"} then the return value is developed from the
\code{file} item in the \code{data} slot of \code{x}, in one of four sub-cases:
\enumerate{
\item If \code{j} is not supplied, the return value is \code{file},
i.e. a vector holding the full names of the files downloaded
by \code{\link[=getProfiles]{getProfiles()}}.
\item Otherwise, if \code{j} is numeric, then the return value is a vector of the
\code{file} entries, as indexed by \code{j}.
\item Otherwise, an error is reported.
}
\item If \code{i} is \code{"length"}, the number of local file names that were downloaded
by \code{\link[=getProfiles]{getProfiles()}} is returned.
}
\item Otherwise, if \code{type} is \code{"argos"}, i.e. if \code{x} was created with \code{\link[=readProfiles]{readProfiles()}}, then:
\enumerate{
\item If \code{i} is numeric and \code{j} is unspecified, then return the argo objects identified
by using \code{i} as an index.  Thus, e.g. \code{x[[i]]} is equivalent to
\code{x[["profile", i]]}.
\item If \code{i} is the name of an item in the \code{metadata} slot, then that item
is returned. There is only choice, \code{"type"}.
\item Otherwise, if \code{i} is the name of an item in the \code{data} slot, then that item
is returned.  There is only one choice: \code{"argos"}.
\item Otherwise, if \code{i} is \code{"profile"} then the return value depends on the value of \code{j}.
There are four sub-cases:
\enumerate{
\item If \code{j} is not supplied, the return value is a list containing
all the profiles in \code{x}, each an \code{argo} object as created by
\code{\link[oce:read.argo]{oce::read.argo()}} in the \code{oce} package.
\item Otherwise, if \code{j} is a single integer,  then the return value is a single
\code{argo} object.
\item Otherwise, if \code{j} is a vector  of integers, then a list of \code{argo} objects
is returned.
\item Otherwise, an error is reported.
}
\item Otherwise, if \code{i} is \code{"length"}, the number of oce-type argo objects in \code{x} is returned.
\item Otherwise, if \code{i} is \code{"ID"}, a vector of the IDs of the profiles is returned.
\item Otherwise, if \code{i} is a character value, then it is taken to be
an item within the \code{metadata} or \code{data} slots of the argo objects
stored in \code{x}, and the returned value is a list containing that
information with one (unnamed) item per profile.  Note that
items from the \code{metadata} are repeated and cast into a matrix
as the same dimensionality as pressure, as a way to make it
easier to e.g. use \code{\link[oce:as.ctd]{oce::as.ctd()}} to create a CTD object from
values returned by the present function, passed through \code{\link[=unlist]{unlist()}}.
\item Otherwise, an error is reported.
}
\item Otherwise, an error is reported.
}
}
\examples{
data(index)
# Full remote filename for first two item in index
paste0(index[["server"]], "/dac/", index[["profile", 1:2]])
# File names and geographical locations of first 5 items in index
index5 <- subset(index, 1:5)
data.frame(file=gsub(".*/", "", index5[["file"]][1]),
           lon=index5[["longitude"]],
           lat=index5[["latitude"]])

}
\author{
Dan Kelley
}
