% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{[[,argoFloats-method}
\alias{[[,argoFloats-method}
\title{Look up a Value Within an argoFloats Object}
\usage{
\S4method{[[}{argoFloats}(x, i, j, ...)
}
\arguments{
\item{x}{an \linkS4class{argoFloats} object.}

\item{i}{a character value that specifies the item to be looked up;
see \dQuote{Details}.}

\item{j}{supplemental index value, ignored unless \code{i} is \code{"profile"} (see \dQuote{Details}).}

\item{...}{ignored.}
}
\description{
This function provides an easy way to look up values within an \linkS4class{argoFloats}
object, without the need to know the exact structure of the data. The action
taken by \code{[[} depends on the \code{type} element in the \code{metadata} slot of
the object (which is set by the function that created the object), on the
value of \code{i} and, in some cases, on the value of \code{j}; see \dQuote{Details}.
}
\details{
\enumerate{
\item If \code{i=="metadata"} then the \code{metadata} slot of \code{x} is returned.
\item If \code{i=="data"} then the \code{data} slot of \code{x} is returned.
\item If \code{i=="ftpRoot"} or \code{"server"} and the object was created with \code{\link[=getIndex]{getIndex()}},
then then the value of that item in the \code{metadata}
slot is returned.  (Note that \code{"server"} is used by \code{\link[=getProfiles]{getProfiles()}} to construct
a vector of URLs to download.)
\item If \code{i=="destdir"} then the \code{destdir} item in the \code{metadata} slot is
returned. This is used by \code{\link[=getProfiles]{getProfiles()}} to decide where to
save downloaded files, and later by \code{\link[=readProfiles]{readProfiles()}}, which
works with the output from \code{\link[=getProfiles]{getProfiles()}}.
\item If \code{i=="type"} then the \code{type} item in the \code{metadata} slot is
returned. This is \code{"index"} if the object was created with
\code{\link[=getIndex]{getIndex()}}, \code{"profiles"} if it was created with \code{\link[=getProfiles]{getProfiles()}},
and \code{"argos"} if it was created with \code{\link[=readProfiles]{readProfiles()}}.
\item If \code{i=="profile"} and \code{x} was created with \code{\link[=readProfiles]{readProfiles()}},
then the return value depends on the value of \code{j}. There are 4
sub-cases.  (a) If \code{j} is not supplied, the return values is a list containing
all the profiles in \code{x}, each an \code{argo} object as created by
\code{\link[oce:read.argo]{oce::read.argo()}} in the \code{oce} package.  (b) If \code{j} is a single integer,
then the return value is a single \code{argo} object. (c) If \code{j} is a vector
of integers, then a list of \code{argo} objects is returned. (d) If \code{j} is
the string \code{"count"}, then the number of profiles is returned.
\item Otherwise, if \code{x} was created with \code{\link[=getIndex]{getIndex()}},
\code{i} may be the name of an item in the \code{index} item
in the \code{data} slot, or a string that is made up of enough characters
to uniquely identify such an item, e.g. \code{"lon"} may be used as a
shortcut for \code{"longitude"}.
}
}
\examples{
data(index)
# Full remote filename for first two item in index
paste0(index[["server"]], "/dac/", index[["profile", 1:2]])
# File names and geographical locations of first 5 items in index
index5 <- subset(index, 1:5)
data.frame(file=gsub(".*/", "", index5[["file"]][1]),
           lon=index5[["longitude"]],
           lat=index5[["latitude"]])

}
\author{
Dan Kelley
}
