% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{[[,argoFloats-method}
\alias{[[,argoFloats-method}
\title{Extract Something From an argoFloats Object}
\usage{
\S4method{[[}{argoFloats}(x, i, j, ...)
}
\arguments{
\item{x}{a \linkS4class{argoFloats} object.}

\item{i}{a character value specifying an item within the object's \code{data} slot. Partial
matches are allowed, e.g. \code{x[["lon"]]} functions the same as \code{x[["lon"]]}, because only
item matches to those three letters.}
}
\description{
Extract Something From an argoFloats Object
}
\author{
Dan Kelley
}
