% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset,argoFloats-method}
\alias{subset,argoFloats-method}
\title{Subset an argoFloats Object}
\usage{
\S4method{subset}{argoFloats}(x, subset = NULL, ...)
}
\arguments{
\item{x}{an \linkS4class{argoFloats} object as created by \code{\link[=getIndex]{getIndex()}}.}

\item{subset}{optional numerical or logical vector that indicates which
indices of \code{x@data$index} to keep.  See example 1.}

\item{...}{a list named \code{circle}, \code{rectangle}, \code{parameter}, \code{polygon} , or \code{time}.
See \dQuote{Details} and \dQuote{Examples}.}
}
\value{
An \linkS4class{argoFloats} object.
}
\description{
Isolate a subset of an \linkS4class{argoFloats} object of type \code{"index"},
as created by \code{\link[=getIndex]{getIndex()}}, either by specifying
indices to keep (using the \code{subset} argument) or by specifying
a way to determine those indices (using the \code{...} argument).
}
\details{
The possibilities for the \code{...} argument are as follows.
\enumerate{
\item A list named \code{circle} with elements named \code{longitude},
\code{latitude} and \code{radius}.  The first two give the center of
the subset region, and the third gives the radius of
that region, in kilometers. See Example 2A.
\item A list named \code{rectangle}, which has an element named \code{longitude}
with two elements giving the western and eastern limits of the
subset region, and similar one named \code{latitude} giving the southern
and northern limits. See Example 2B.
\item A list named \code{polygon} that has elements named \code{longitude} and \code{latitude}.
Profiles within this polygon will be retained. See Example 2C.
\item A vector or list named \code{parameter} that holds names of parameters to keep;
see examples 3A and 3B.
\item A list named \code{time} that has elements \code{from} and \code{to} that are POSIXt
times created with eg. \code{\link[=POSIXct]{POSIXct()}}, with \code{tz="UTC"} to match
the timezone used in Argo data. Profiles within that time frame will
be retained.
\item A list named \code{institution}, which has \code{argo}, \code{argo_bgc}, or \code{argo_merge}
institutions (ie. \code{AO} meaning AOML, USA, \code{BO} meaning BODC, United Kingdom,
\code{CS} meaning CSIRO, Austrialia, \code{HZ} meaning CSIO, China Second Institute of
Oceanography, \code{IF} meaning Ifremer, France, \code{IN} meaning INCOIS, India, \code{JA}
meaning JMA, Japan, \code{KM} meaning KMA, Korea, \code{KO} meaning KORDI, Korea, \code{ME}
meaning MEDS, Canada, and \code{NM} meaning NMDIS, China).
\item A list named \code{deep}, which has a logical value indicating weather argo floats
are deep argo (ie. `profiler_type`` 849, 862, and 864).
\item A list named \code{ID}, which has \code{argo}, \code{argo_bgc}, or \code{argo_merge} float ID
numbers.
}

In all cases, the notation is that longitude is positive
for degrees East and negative for degrees West, and that latitude
is positive for degrees North and negative for degrees South.
}
\examples{
library(argoFloats)
data(index)

# Example 1: First three profiles in dataset.
index1 <- subset(index, 1:3)
cat("First 3 longitudes:", paste(index1[["longitude"]]), "\n")

# Example 2: Demonstrate geographical subsets. These examples use
# built-in datasets; to use these in other  work, start by
# using getIndex() to create an index.
# 2A: circle north of Abaca Island
index2A <- subset(index, circle=list(longitude=-77.5, latitude=27.5, radius=50))
# 2B: rectangle northeast of Abaca Island
lonRect <- c(-76.5, -76)
latRect <- c(26.5, 27.5)
index2B <- subset(index, rectangle=list(longitude=lonRect, latitude=latRect))
# 2C: polygon to the southeast of Abaco Island
lonPoly <- c(-76.5, -76.0, -75.5)
latPoly <- c(25.5, 26.5, 25.5)
index2C <- subset(index, polygon=list(longitude=lonPoly, latitude=latPoly))
# Show these subsets on a map
plot(index)
points(index2A[["longitude"]], index2A[["latitude"]], col=2, pch=20, cex=1.4)
points(index2B[["longitude"]], index2B[["latitude"]], col=3, pch=20, cex=1.4)
rect(lonRect[1], latRect[1], lonRect[2], latRect[2], border=3, lwd=2)
points(index2C[["longitude"]], index2C[["latitude"]], col=4, pch=20, cex=1.4)
polygon(lonPoly, latPoly, border=4, lwd=2)

# Example 3: Subsetting argo_merge data containing 'DOXY' parameters
# 3A: Data containing 'DOXY' data
data(indexMerged)
index3A <- subset(indexMerged, parameter="DOXY")
# 3B: Data containing both 'PSAL' and 'DOWN_IRRADIANCE380' data
data(indexMerged)
index3B <- subset(indexMerged, parameter=c("PSAL", "DOWN_IRRADIANCE380"))

# Example 4: Subsetting data for the year 2019
data(index)
from <- as.POSIXct("2019-01-01", tz="UTC")
to <- as.POSIXct("2019-12-31", tz="UTC")
index4 <- subset(index, time=list(from=from, to=to))

# Example 5: Subsetting data to only include deep argo
\dontrun{
ai <- getIndex(file='merged', destdir = '~/data/argo')
index5 <- subset(ai, deep=TRUE)
summary(index5) }

# Example 6: Subsetting by specific argo ID
\dontrun{
ai <- getIndex(file='merged', destdir = '~/data/argo')
index6 <- subset(ai, ID='1900722')
summary(index6) }

}
\author{
Dan Kelley and Jaimie Harbin
}
