% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{subset,argoFloats-method}
\alias{subset,argoFloats-method}
\alias{subset.argoFloats}
\title{Subset an argoFloats Object}
\usage{
\S4method{subset}{argoFloats}(x, subset = NULL, ...)
}
\arguments{
\item{x}{an \linkS4class{argoFloats} object.}

\item{subset}{optional numerical or logical vector that indicates which
indices of \code{x@data$index} to keep.  See example 1.}

\item{...}{a single named list. At the moment, the only possibility
is a list named \code{circle} and \code{rectangle}. \code{circle} holds elements named \code{longitude},
\code{latitude} and \code{radius} . The first two specify a location, and
the third species a search radius in kilometers, as in Example 2.
\code{rectangle} holds elements named \code{longitude} and \code{latitude} as in Example 3.}
}
\value{
An \linkS4class{argoFloats} object.
}
\description{
Return a subset of an \linkS4class{argoFloats} object, either by specifying
indices to keep (using the \code{subset} argument) or by specifying
a way to determine those indices (using the \code{...} argument).
}
\details{
At the moment, this method only works for \linkS4class{argoFloats} objects
of type \code{"index"}, as read by \code{\link[=getIndex]{getIndex()}}.
A future version will also handle \code{...} arguments named
\code{polygon}, and \code{time}.
}
\examples{
library(argoFloats)
\dontrun{
ai <- getIndex(file="argo_bio-profile_index.txt.gz", destdir="~/data/argo")
# Example 1: First three profiles in dataset.
aiFirstThree <- subset(ai, 1:3)
cat("First three longitudes:", paste(aiFirstThree[["longitude"]]), "\n")

# Example 2: Profiles within 200km of Sable Island
aiSable <- subset(ai, circle=list(longitude=-59.915, latitude=44.934, radius=200))
cat("Found", length(aiSable[["longitude"]]), "profiles near Sable Island\n")
}

# Example 3: Profiles in a given rectangle radius

## Not run:
subset(ai, rectangle=list(longitude=c(-65,-64), latitude=c(40,45)))

}
\author{
Dan Kelley
}
