% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,argoFloats-method}
\alias{summary,argoFloats-method}
\title{Summarize an argoFloats Object}
\usage{
\S4method{summary}{argoFloats}(object, ...)
}
\arguments{
\item{object}{an \code{\linkS4class{argoFloats}} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
None (invisible NULL).
}
\description{
Show some key facts about an \code{\linkS4class{argoFloats}} object.
}
\examples{
library(argoFloats)
data(index)
summary(index)

}
\author{
Dan Kelley
}
