% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{useAdjusted}
\alias{useAdjusted}
\title{Switch [[ and Plot to Focus on Adjusted data}
\usage{
useAdjusted(argos, fallback = "NA", debug = 0)
}
\arguments{
\item{argos}{an \code{\linkS4class{argoFloats}} object, as read by \code{\link[=readProfiles]{readProfiles()}}.}

\item{fallback}{a character value indicating what to do if all the adjusted
values for a particular parameter-profile pair are \code{NA}. The choices are
\code{"NA"} and \code{"raw"}; see \dQuote{Details}.}

\item{debug}{an integer that controls whether debugging information is printed
during processing.  If \code{debug} is 0 or less, then no information is printed.
If it is 1 then minimal overall information is printed.  If it exceeds 1,
then information is printed for each Argo cycle contained in \code{argos}.}
}
\description{
This function returns a copy of its first argument that has been modified by
(optionally) replacing 'raw' parameter values, and their flags, with
corresponding 'adjusted' values. The purpose is to ensure that future calls
to \code{\link{[[,argoFloats-method}} and plotting methods will focus on
the adjusted data instead of the raw data.  The procedure hinges on the value
of the \code{fallback} argument.
}
\details{
There are two cases, depending on the value of \code{fallback}.
\enumerate{
\item If \code{fallback} is \code{"NA"} (the default), then
the adjusted values become the focus, even if they are all \code{NA}. Thus,
\code{fallback="NA"} may be preferred as part of a cautious analysis that
focuses only on data of high quality. The downside of this approach is a
reduction in coverage, since it eliminates the 'raw' fields that are typical
of real-time mode data components (see  Section 2.2.5 of Reference 1).
\item If \code{fallback} is \code{"raw"}, then the data in the adjusted parameter
data columns (corresponding to profiles) of each cycle are examined one by
one. If the values in a given column are all \code{NA}, then the parameter's raw
values for that column are not altered. However, if any of the values in a
given adjusted column are non-\code{NA}, then that entire 'adjusted' column is
copied into the corresponding 'raw' column. Since the \code{fallback="raw"}
scheme combines 'raw' and 'adjusted' data, it increases coverage, at the
possible expense of lowering overall data quality.
}
}
\examples{
library(argoFloats)
file <- "SD5903586_001.nc"
A <- readProfiles(system.file("extdata", file, package="argoFloats"))
B <- useAdjusted(A, "NA")
C <- useAdjusted(A, "raw")
# Notice that the original values are smaller than the adjusted values
plot(C, which="profile", profileControl=list(parameter="oxygen"), pch=1)
points(A[[1]][["oxygen"]], A[[1]][["pressure"]], pch=2)
points(B[[1]][["oxygen"]], B[[1]][["pressure"]], pch=3)
legend("bottomright", pch=c(3,1,2), legend=c("fallback='NA'", "fallback='raw'", "original"))

}
\references{
\enumerate{
\item Argo Data Management Team. "Argo User’s Manual V3.4,"
January 20, 2021. \url{https://archimer.ifremer.fr/doc/00187/29825/}
}
}
\author{
Dan Kelley and Jaimie Harbin
}
