% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{useAdjusted}
\alias{useAdjusted}
\title{Switch [[ and Plot to Focus on Adjusted data}
\usage{
useAdjusted(argos, fallback = FALSE, debug = 0)
}
\arguments{
\item{argos}{an \code{\linkS4class{argoFloats}} object, as read by \code{\link[=readProfiles]{readProfiles()}}.}

\item{fallback}{a logical value indicating whether to 'fall back' from
adjusted data to raw data for profiles that are in real-time mode.  By
default, \code{fallback} is \code{FALSE}, to focus entirely on
adjusted data.  See \dQuote{Details}.}

\item{debug}{an integer that controls whether debugging information is printed
during processing.  If \code{debug} is 0 or less, then no information is printed.
If it is 1 then minimal overall information is printed.  If it exceeds 1,
then information is printed for each Argo cycle contained in \code{argos}.}
}
\description{
\code{useAdjusted} returns a copy of an \code{argos}-type object (as created with the
\code{\link[=readProfiles]{readProfiles()}} function), in which the individual
\link[oce:argo-class]{oce::argo} objects have been modified so that future calls to
\verb{[[,argoFloats-method} or the \verb{plot,argoFloats-method()} will work with
\emph{adjusted} versions of the data.  In effect, this changes the focus from the
observed data to adjusted data.  This works profile-by-profile, for each of
the \link[oce:argo-class]{oce::argo} objects stored within the first argument.
The details of how this works are controlled by
the \code{fallback} argument.
}
\details{
If \code{fallback} is \code{FALSE}, which the default, then the focus switches entirely
to the adjusted data.  This improves the overall reliability of the results,
but at the cost of eliminating realtime-mode data.  This is because the
adjusted fields for realtime data are set to \code{NA} as a matter of policy (see
section 2.2.5 of reference 1).

Wider data coverage is obtained if \code{fallback} is set to \code{TRUE}.  In this
case, the focus is shifted to adjusted data \emph{only if} the data-mode for
the individual profiles is \code{A} or \code{D}, indicating either Adjusted or
Delayed mode. Any profiles that are of the \code{R} (Realtime) data-mode are
left unaltered. This blending of adjusted and unadjusted data offers
improved spatial and temporal coverage, while reducing the overall
data quality, and so this approach should be used with caution.
}
\examples{
library(argoFloats)
file <- "SD5903586_001.nc"
A <- readProfiles(system.file("extdata", file, package="argoFloats"))
B <- useAdjusted(A)
C <- useAdjusted(A, TRUE)
# Notice that the original values are smaller than the adjusted values
plot(C, which="profile", profileControl=list(parameter="oxygen"), pch=1)
points(A[[1]][["oxygen"]], A[[1]][["pressure"]], pch=2)
points(B[[1]][["oxygen"]], B[[1]][["pressure"]], pch=3)
legend("bottomright", pch=c(3,1,2),
       legend=c("fallback=TRUE", "fallback=FALSE", "original"))

}
\references{
\enumerate{
\item Argo Data Management Team. "Argo User’s Manual V3.4,"
January 20, 2021. \url{https://archimer.ifremer.fr/doc/00187/29825/}
}
}
\author{
Dan Kelley, Jaimie Harbin and Clark Richards
}
