% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{useAdjusted}
\alias{useAdjusted}
\title{Switch unadjusted and adjusted data}
\usage{
useAdjusted(argo, debug = 0)
}
\arguments{
\item{argo}{An \code{\linkS4class{argoFloats}} object, as read by \code{\link[=readProfiles]{readProfiles()}}.}

\item{debug}{An integer that, if positive, indicates that some debugging information
should be printed.}
}
\description{
Variables with original names indicating in the string \verb{_ADJUSTED} are
assigned nicknams names ending in \code{Adjusted} by \code{\link[=readProfiles]{readProfiles()}}, so that
e.g. \code{TEMP_ADJUSTED} gets the nickname \code{temperatureAdjusted}, while
\code{TEMP} gets the nickname \code{temperature}.  \code{\link[=useAdjusted]{useAdjusted()}} switches
these, renaming the adjusted values, so that e.g. \code{TEMP_ADJUSTED}
gets nickname \code{temperature} and \code{TEMP} gets nickname \code{temperatureUnajusted}.
This is carried out not just for data, and also for corresponding units and
quality-control flags.
}
\author{
Dan Kelley
}
