% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{useAdjusted}
\alias{useAdjusted}
\title{Switch unadjusted and adjusted data}
\usage{
useAdjusted(argo, debug = 0)
}
\arguments{
\item{argo}{An \code{\linkS4class{argoFloats}} object, as read by \code{\link[=readProfiles]{readProfiles()}}.}

\item{debug}{An integer that, if positive, indicates that some debugging information
should be printed.}
}
\description{
This is done by using \code{\link[oce:preferAdjusted]{oce::preferAdjusted()}} on each of the Argo
objects stored within the first argument.
}
\details{
This is a complex process, however, simply put, \code{useAdjusted()} allows users
to express preference for adjusted vs unadjusted data. If an Adjusted field exists,
and it is not filled with all NA values, this function will return the Adjusted
data. If there is no Adjusted data, or it is all NA values, \code{useAdjusted()} will
return the original data.

See Reference 1 for the process of adding adjusted data to Argo
files, and see the documentation of \code{\link[oce:preferAdjusted]{oce::preferAdjusted()}} for the
details of how the preference for adjusted data is set.
}
\examples{
library(argoFloats)
# Plot raw and adjusted oxygen for a sample dataset
raw <- readProfiles(system.file("extdata", "SD5903586_001.nc", package="argoFloats"))
adj <- useAdjusted(raw)
par(mfrow=c(1, 2))
oce::plotProfile(oce::as.ctd(raw[[1]]), xtype="oxygen")
mtext("Raw data", side=3, line=-1, col=2)
oce::plotProfile(oce::as.ctd(adj[[1]]), xtype="oxygen")
mtext("Adjusted data", side=3, line=-1, col=2)

}
\references{
Carval, Thierry, Bob Keeley, Yasushi Takatsuki, Takashi Yoshida, Stephen Loch Loch,
Claudia Schmid, and Roger Goldsmith. Argo User’s Manual V3.3. Ifremer, 2019.
\url{https://doi.org/10.13155/29825}.
}
\author{
Dan Kelley and Jaimie Harbin
}
